Renaming relaxed_constants, adding placeholder test.
