+++ gpkit/tests/run_tests.py
@@ -39,8 +39,10 @@ def import_tests():
     from gpkit.tests import t_keydict
     tests += t_keydict.TESTS
 
-    return tests
+    from gpkit.tests import t_algorithms
+    tests += t_algorithms.TESTS
 
+    return tests
 
 def run(xmloutput=False, tests=None, verbosity=1):
     """Run all gpkit unit tests.
